From: Willem <willem@king-pin.nl> To: ezmlm@list.cr.yp.to Subject: Dutch version of ezmlmrc Date: Mon, 4 Aug 2003 17:18:10 +0200 Message-Id: <200308041718.10550.willem@king-pin.nl>
Hi,

It seemed that the ezmlmrc file was translated to every language but Dutch.
Here is a crude version as attachment.
Some things don't translate very well from English to Dutch, they get an
undertone of disrespect towards the reader. So, if you can improve it, please
do and be so kind to send me a copy.

Regards,
Willem
the Netherlands