Date: Wed, 14 Mar 2001 22:25:36 -0800 From: edk@collab.net To: ezmlm@list.cr.yp.to, Fred Lindberg <lindberg@id.wustl.edu> Subject: PATCH : ezmlm-send.c Message-ID: <20010314222536.O27952@collab.net>
The attached patch fixes a problem with ezmlm-send in ezmlm-idx (up to
0.40 at least): it munges MIME messages which use the default type
implicitly in the first MIME part.  (This is not an issue for stock
ezmlm, which doesn't attempt to parse MIME parts AFAICT.)  The patch is
applied after all the idx stuff is; I think it should be integrated into
the -idx patches.  (Copyright is granted, of course.)

See http://www.netbeans.org/issues/show_bug.cgi?id=9039 for a
description of the bug this fixes.

I'll be happy to answer any questions ...

thanks --

Ed