Message-ID: <20000109163828.632.qmail@id.wustl.edu> From: "Fred Lindberg" <lindberg@id.wustl.edu> To: "ezmlm@list.cr.yp.to" <ezmlm@list.cr.yp.to> Date: Sun, 09 Jan 2000 10:36:39 -0600 Subject: Re: Subscriber-only, moderated lists don't get ezmlm-archived?
Just insert

|/.../ezmlm-archive DIR || exit 0

into DIR/editor after ezmlm-send (in addition to the DIR/moderator
invocation where it already should be for ium lists).

Or edit ezmlmrc (see below), then ezmlm-make -+ DIR

That special case isn't covered for ezmlm-archive, which of course it
should be. For moderated messages it works, but ezmlm-archive is not
invoked by messages from subscribers (so they won't be entered into the
index until the next moderated message comes in).