Fixed ezmlm-send to insert the proper value for the List-ID: header.
From: Bill Nugent <whn@lopi.com>
To: ezmlm@list.cr.yp.to
Subject: ezmlm-idx-5.1.1 bug in ezmlm-send.c
Date: Fri, 23 Feb 2007 10:04:08 -0500
Message-Id: <200702231004.08420.whn@lopi.com>

Howdy,

Running ezmlm-0.53 patched with ezmlm-idx 5.1.1 came across a problem
where the headers were trashed/garbled  when dir/listid exists.  Search
a list archive to no luck on this problem.  So dug into the source code
and found what appears to be a bug on line 481 where the wrong variable
is being appended to the List-ID: header line and changing it gives the
expected results.  Here is a patch:

$ diff -u ezmlm-send.c.orig ezmlm-send.c
--- ezmlm-send.c.orig   2007-02-23 14:54:43.437831108 +0000
+++ ezmlm-send.c        2007-02-23 14:46:01.412832612 +0000
@@ -478,7 +478,7 @@
     if (listid.len > 0) {
       flaglistid = 1;
       qmail_puts(&qq,"\nList-ID: ");
-      qmail_put(&qq,line.s,line.len);
+      qmail_put(&qq,listid.s,listid.len);
     }
     qa_puts("\n");
   }

I have not been a subscriber to the list for several years so if there
are any responses please include me in the cc.   BTW - 5.1.1 is a nice
improvement over ezmlm-idx 0.40.  It is only getting better.  Thank
you.

	Bill