Modified the handling of -l libraries in *.do
When a -l* line is present in a *=x or *=so file, normally lib*.a is
added to the list of dependancies.  This change first checks if such a
library can be built (by looking for the corresponding *=l file) before
adding the dependancy.  This allows -l lines that specify external
required system libraries.