Use the message name as the key instead of the text.
The problem with the previous approach shows up when the original
message text is altered.  With the previous approach, the result will be
that the original message will be output instead of the older translated
message.  This differs from the handling of the other text files, and
would likely be considered surprising.

Using the message name as the key also makes for shorter and easier to
understand messages files.