Generate lang/*/text/messages from new lang/*/messages files.
The source files contain lines of the form:

	NAME:ORIGINAL:TRANSLATION

Where "NAME" is the message name, "ORIGINAL" is the original text, and
"TRANSLATION" is the current translation.  This will allow translation
tools to determine if the translation for a given name is out of date
without preventing ezmlm-idx from outputting a translated message.