Make the -h option the default for ezmlm-reject.
Since the default ezmlmrc doesn't create a headerreject file, and lists
with one present will normally have set the -h flag, this change should
not adversely affect anyone.  It will make it easier to set up the
headerreject feature.

With this change, a missing headerreject file is no longer treated as an
error, also to prevent compatibility issues.