If "flags" does not exist, try to load flags from "config"
On lists containing only a "config" file (and no "flags" or
"addtrailer") files, the trailer would never be added to outgoing
messages.  This fixes handling of adding the trailer on such old lists,
since new ezmlm ignores the presence of a "text/trailer" as a flag for
adding the trailer.