getconfopt: Read default files before parsing command line values
This fixes a problem found particularly in ezmlm-store but may exist in
other programs.  There, there are a pair of flag options used to set a
single flag variable, the first of which referenced a default file.  If
the latter (with no default file) was specified on the command line, the
default would still be loaded from the file and would reset the flag.

The core problem is that the parsing relied on an "is set" flag to
determine if the default file should be consulted, but that flag was
stored with the option character specification rather than with the
option value itself.  Where there would be a pair of characters used to
set the same flag, one could override the other, since only one "is set"
flag would be true.

This is the fourth approach tried.  The first was simply to give flag
options on+off characters, but some flags are used with values other
than 1/0 so that failed.  The other two involved creating new data types
for the option values, but they would involve rewriting most code that
handled such options, and so were too intrusive.  This one simply
reverses the order of looking up default files and parsing the command
line, making the "is set" flag irrelevant.

There is one gotcha, which makes the patch larger than otherwise
necessary, and that is that the files must be checked *after* changing
to the list directory, but the list directory is on the command line
after the options.  So, two passes are made through the command line,
the first to find the directory option, then chdir and read the files,
and finally actually parse the command line options.