unfoldHDR: Fix missing pointer increment and var initialization
The unfoldHDR function was missing a pointer increment when handling
iso-2022-{cn,kr} character sets, preventing them from being used
properly.  As far as I can tell, this bug was introduced prior to
version 0.42.

Also, the cset variable was potentially used unitialized in the same
function.  This problem was only revealed when the unreachable code
caused by the above bug was corrected.