Handle From header lacking whitespace after colon
When rewritefrom needs to append the original From address to an existing Cc
header the from address is placed on the next line. Whitespace is required
at the beginning of the line so the added line is treated as an extension of
the previous header. Ensure there is whitespace at the beginning of the
line, in case there was not any after the colon in the From header.